refactor: use direct imports for internal utilities
